--- .gitmessage
@@ -1,8 +1,0 @@
-What is new: 
-
-
-Authors: 
-Roman Zvoda - rozv@itu.dk
-
-
-ChatGPT
