--- Makefile
@@ -1,8 +1,0 @@
-init:
-	python -c"from minitwit import init_db; init_db()"
-
-build:
-	gcc flag_tool.c -l sqlite3 -o flag_tool
-
-clean:
-	rm flag_tool
