+++ README.md
@@ -1,0 +1,11 @@
+# Welcome to the new MiniTwit app!
+
+## Folders
+
+
+*Authors:*
+Daniel Grønbjerg
+Roman Zvoda
+Viktoria Hopeberg
+Trond Rossing
+Rasmus Balder Nordbjærg
