+++ docs/working-process.md
@@ -1,0 +1,16 @@
+# Ways of working
+
+## Change process
+1. Create branch based of main.
+2. Implement code changes.
+3. Create PR.
+4. Have another person review it.
+5. Reviewer should understand and have tested the code before approving.
+6. Merge PR using "squash commits" option.
+
+## Commit guidelines
+* Keep commits small and concise.
+* Try to have on change per PR working on 2 separate issues should be 2 seperate PRs.
+* No reverts.
+* If the branch has not been merged with main for a long time, then do a local merge before making a PR.
+* If broadly spanding changes are required this needs to be aligned with the rest of the team.
