--- go.mod
@@ -1,11 +1,0 @@
-module MiniTwit
-
-go 1.22.0
-
-require (
-	github.com/gorilla/mux v1.8.1 // indirect
-	github.com/gorilla/securecookie v1.1.2 // indirect
-	github.com/gorilla/sessions v1.2.2 // indirect
-	github.com/itchyny/timefmt-go v0.1.5 // indirect
-	github.com/mattn/go-sqlite3 v1.14.22 // indirect
-)
