--- go/templates/login_test.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Test</title>
-  <link rel="stylesheet" type="text/css" href="/static/style.css">
-</head>
-<body>
-  <h2>Sign In {{ .Name }} - test</h2>
-  <form action="" method="post">
-    <dl>
-      <dt>Username:</dt>
-      <dd><input type="text" name="username" size="30"></dd>
-      <dt>Password:</dt>
-      <dd><input type="password" name="password" size="30"></dd>
-    </dl>
-    <div class="actions"><input type="submit" value="Sign In"></div>
-  </form>
-</body>
-</html>
