--- old_app/Dockerfile
@@ -1,6 +1,0 @@
-FROM python:latest
-
-RUN pip install Werkzeug
-RUN pip install flask
-COPY . .
-CMD [ "python", "./minitwit.py" ]
