--- old_app/static/style.css
@@ -1,178 +1,0 @@
-body {
-    background: #CAECE9;
-    font-family: 'Trebuchet MS', sans-serif;
-    font-size: 14px;
-}
-
-a {
-    color: #26776F;
-}
-
-a:hover {
-    color: #333;
-}
-
-input[type="text"],
-input[type="password"] {
-    background: white;
-    border: 1px solid #BFE6E2;
-    padding: 2px;
-    font-family: 'Trebuchet MS', sans-serif;
-    font-size: 14px;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    color: #105751;
-}
-
-input[type="submit"] {
-    background: #105751;
-    border: 1px solid #073B36;
-    padding: 1px 3px;
-    font-family: 'Trebuchet MS', sans-serif;
-    font-size: 14px;
-    font-weight: bold;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    color: white;
-}
-
-div.page {
-    background: white;
-    border: 1px solid #6ECCC4;
-    width: 700px;
-    margin: 30px auto;
-}
-
-div.page h1 {
-    background: #6ECCC4;
-    margin: 0;
-    padding: 10px 14px;
-    color: white;
-    letter-spacing: 1px;
-    text-shadow: 0 0 3px #24776F;
-    font-weight: normal;
-}
-
-div.page div.navigation {
-    background: #DEE9E8;
-    padding: 4px 10px;
-    border-top: 1px solid #ccc;
-    border-bottom: 1px solid #eee;
-    color: #888;
-    font-size: 12px;
-    letter-spacing: 0.5px;
-}
-
-div.page div.navigation a {
-    color: #444;
-    font-weight: bold;
-}
-
-div.page h2 {
-    margin: 0 0 15px 0;
-    color: #105751;
-    text-shadow: 0 1px 2px #ccc;
-}
-
-div.page div.body {
-    padding: 10px;
-}
-
-div.page div.footer {
-    background: #eee;
-    color: #888;
-    padding: 5px 10px;
-    font-size: 12px;
-}
-
-div.page div.followstatus {
-    border: 1px solid #ccc;
-    background: #E3EBEA;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    padding: 3px;
-    font-size: 13px;
-}
-
-div.page ul.messages {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-}
-
-div.page ul.messages li {
-    margin: 10px 0;
-    padding: 5px;
-    background: #F0FAF9;
-    border: 1px solid #DBF3F1;
-    -moz-border-radius: 5px;
-    -webkit-border-radius: 5px;
-    min-height: 48px;
-}
-
-div.page ul.messages p {
-    margin: 0;
-}
-
-div.page ul.messages li img {
-    float: left;
-    padding: 0 10px 0 0;
-}
-
-div.page ul.messages li small {
-    font-size: 0.9em;
-    color: #888;
-}
-
-div.page div.twitbox {
-    margin: 10px 0;
-    padding: 5px;
-    background: #F0FAF9;
-    border: 1px solid #94E2DA;
-    -moz-border-radius: 5px;
-    -webkit-border-radius: 5px;
-}
-
-div.page div.twitbox h3 {
-    margin: 0;
-    font-size: 1em;
-    color: #2C7E76;
-}
-
-div.page div.twitbox p {
-    margin: 0;
-}
-
-div.page div.twitbox input[type="text"] {
-    width: 585px;
-}
-
-div.page div.twitbox input[type="submit"] {
-    width: 70px;
-    margin-left: 5px;
-}
-
-ul.flashes {
-    list-style: none;
-    margin: 10px 10px 0 10px;
-    padding: 0;
-}
-
-ul.flashes li {
-    background: #B9F3ED;
-    border: 1px solid #81CEC6;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    padding: 4px;
-    font-size: 13px;
-}
-
-div.error {
-    margin: 10px 0;
-    background: #FAE4E4;
-    border: 1px solid #DD6F6F;
-    -moz-border-radius: 2px;
-    -webkit-border-radius: 2px;
-    padding: 4px;
-    font-size: 13px;
-}
