--- old_app/templates/login.html
@@ -1,16 +1,0 @@
-{% extends "layout.html" %}
-{% block title %}Sign In{% endblock %}
-{% block body %}
-  <h2>Sign In</h2>
-  {% if error %}<div class=error><strong>Error:</strong> {{ error }}</div>{% endif %}
-  <form action="" method=post>
-    <dl>
-      <dt>Username:
-      <dd><input type=text name=username size=30 value="{{ request.form.username }}">
-      <dt>Password:
-      <dd><input type=password name=password size=30>
-    </dl>
-    <div class=actions><input type=submit value="Sign In"></div>
-  </form>
-{% endblock %}
-
