+++ tests/go.mod
@@ -1,0 +1,18 @@
+module minitwit
+
+go 1.18
+
+require (
+	github.com/gorilla/mux v1.8.1
+	github.com/gorilla/sessions v1.2.2
+	github.com/mattn/go-sqlite3 v1.14.22
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gorilla/securecookie v1.1.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.8.4 // indirect
+	golang.org/x/crypto v0.19.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
